Move proposal computers to seperate package, pull out AST visitor to seperate class, add icon.
