+++ org.mapstruct.eclipse/build.properties
@@ -20,4 +20,5 @@
 source.. = src/
 bin.includes = .,\
                META-INF/,\
-               plugin.xml
+               plugin.xml,\
+               icons/
