+++ org.mapstruct.eclipse/src/org/mapstruct/eclipse/internal/util/Bindings.java
@@ -16,7 +16,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.mapstruct.eclipse.internal;
+package org.mapstruct.eclipse.internal.util;
 
 import java.util.ArrayList;
 import java.util.Collection;
